Delete SQL query

+++ OOP/SQL.cpp
@@ -19,9 +19,9 @@ using namespace std;
 wchar_t* char2wchar(const char* cchar)
 {
 	wchar_t* m_wchar;
-	int len = MultiByteToWideChar(CP_ACP, 0, cchar, strlen(cchar), NULL, 0);
+	int len = MultiByteToWideChar(CP_ACP, 0, cchar, (int)strlen(cchar), NULL, 0);
 	m_wchar = new wchar_t[len + 1];
-	MultiByteToWideChar(CP_ACP, 0, cchar, strlen(cchar), m_wchar, len);
+	MultiByteToWideChar(CP_ACP, 0, cchar, (int)strlen(cchar), m_wchar, len);
 	m_wchar[len] = '\0';
 	return m_wchar;
 }
